+++ src/main/java/web/mvc/domain/UserBuy.java
@@ -17,7 +17,7 @@ public class UserBuy {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "buy_seq")
-    private Integer buySeq;
+    private Long buySeq;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_seq", nullable = false)
