+++ src/main/java/web/mvc/repository/StockGradeRepository.java
@@ -3,5 +3,5 @@ package web.mvc.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.mvc.domain.StockGrade;
 
-public interface StockGradeRepository extends JpaRepository<StockGrade, Integer> {
+public interface StockGradeRepository extends JpaRepository<StockGrade, Long> {
 }
