+++ src/main/java/web/mvc/repository/StockOrganicRepository.java
@@ -3,5 +3,5 @@ package web.mvc.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.mvc.domain.StockOrganic;
 
-public interface StockOrganicRepository extends JpaRepository<StockOrganic, Integer> {
+public interface StockOrganicRepository extends JpaRepository<StockOrganic, Long> {
 }
